requirements_vqa.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
allennlp==0.8.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==1.6.6
async-timeout==4.0.2
asynctest==0.13.0
attrdict==2.0.1
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
aws-xray-sdk==0.95
awscli==1.23.1
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.11.1
bleach==5.0.0
blis @ file:///tmp/build/80754af9/cython-blis_1613319337621/work
boto==2.49.0
boto3==1.22.1
botocore==1.25.1
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028896657/work
brotlipy==0.7.0
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
cachetools==5.0.0
catalogue==1.0.0
certifi==2021.10.8
cffi==1.11.5
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056591685/work
codecov==2.1.12
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
commentjson==0.9.0
conllu==0.11
cookies==2.2.1
coverage==6.3.2
cryptography==3.2.1
cycler==0.11.0
cymem @ file:///tmp/build/80754af9/cymem_1613319258118/work
Cython @ file:///tmp/build/80754af9/cython_1647850341668/work
cytoolz==0.9.0.1
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
dill==0.2.9
docker==5.0.3
docutils==0.15.2
ecdsa==0.17.0
editdistance==0.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz
entrypoints==0.4
fastjsonschema==2.15.3
filelock==3.6.0
flaky==3.7.0
Flask==1.0.2
Flask-Cors==3.0.9
frozenlist==1.3.0
ftfy==6.1.1
future==0.18.2
gevent==1.3.6
google-auth==2.6.6
google-auth-oauthlib==0.4.6
graphviz==0.20
greenlet==1.1.2
grpcio==1.44.0
h5py @ file:///tmp/build/80754af9/h5py_1637138881647/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imagesize @ file:///tmp/build/80754af9/imagesize_1637939814114/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
iniconfig==1.1.1
ipykernel==4.10.1
ipython @ file:///tmp/build/80754af9/ipython_1643818144432/work
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
itsdangerous==2.1.2
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
jmespath==1.0.0
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jsondiff==1.1.1
jsonnet==0.10.0
jsonpickle==2.1.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lark-parser==0.7.8
lazy-object-proxy==1.7.1
livereload==2.6.3
Markdown==3.3.6
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib==2.2.3
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mccabe==0.7.0
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mock==4.0.3
moto==1.3.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multidict==6.0.2
murmurhash @ file:///tmp/build/80754af9/murmurhash_1607456104392/work
mypy==0.521
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
nltk==3.6.3
notebook==6.4.11
numexpr @ file:///tmp/build/80754af9/numexpr_1640704205123/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1649782762302/work
numpydoc==0.8.0
oauthlib==3.2.0
opencv-python==4.5.5.64
overrides==6.1.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
pandocfilters==1.5.0
parsimonious==0.8.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
pkginfo==1.8.2
plac==0.9.6
pluggy==1.0.0
preshed==2.0.1
prometheus-client==0.14.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py==1.11.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.14.1
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pylint==1.8.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pypandoc==1.7.5
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636098896055/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-jose==2.0.2
pytorch-pretrained-bert==0.3.0
pytz==2017.3
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2018.1.10
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
responses==0.20.0
rsa==4.7.2
s3transfer==0.5.2
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1642601761909/work
scipy @ file:///tmp/build/80754af9/scipy_1641536880743/work
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
snowballstemmer @ file:///tmp/build/80754af9/snowballstemmer_1637937080595/work
soupsieve==2.3.2.post1
spacy==2.0.18
Sphinx==1.5.3
sphinx-autobuild==2021.3.14
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp @ file:///home/ktietz/src/ci/sphinxcontrib-applehelp_1611920841464/work
sphinxcontrib-devhelp @ file:///home/ktietz/src/ci/sphinxcontrib-devhelp_1611920923094/work
sphinxcontrib-htmlhelp @ file:///tmp/build/80754af9/sphinxcontrib-htmlhelp_1623945626792/work
sphinxcontrib-jsmath @ file:///home/ktietz/src/ci/sphinxcontrib-jsmath_1611920942228/work
sphinxcontrib-qthelp @ file:///home/ktietz/src/ci/sphinxcontrib-qthelp_1611921055322/work
sphinxcontrib-serializinghtml @ file:///tmp/build/80754af9/sphinxcontrib-serializinghtml_1624451540180/work
sqlparse==0.2.4
srsly @ file:///tmp/build/80754af9/srsly_1607548514031/work
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==1.2
terminado==0.13.3
thinc==6.12.1
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tinycss2==1.1.1
tomli==2.0.1
toolz==0.11.2
torch==1.11.0
torchaudio==0.11.0
torchsummary==1.5.1
torchvision==0.12.0
torchviz==0.0.2
tornado==6.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1650891076910/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
twine==1.11.0
typed-ast==1.0.4
typing-utils==0.1.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
ujson==5.2.0
Unidecode==1.3.4
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
wasabi @ file:///tmp/build/80754af9/wasabi_1612219178408/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
wget==3.2
widgetsnbextension==3.6.0
wrapt==1.10.11
xmltodict==0.12.0
yarl==1.7.2
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
